cast_service/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
databases==0.8.0
exceptiongroup==1.1.3
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
databases[postgresql]==0.8.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
psycopg2-binary==2.9.7
psycopg2==2.9.7